build fix.
svn:sync-xref-src-repo-rev: r99341

+++ src/VBox/VMM/testcase/tstHelp.h
@@ -1,4 +1,4 @@
-/* $Id: tstHelp.h 55051 2015-03-31 19:06:19Z knut.osmundsen@oracle.com $ */
+/* $Id: tstHelp.h 55052 2015-03-31 19:09:50Z knut.osmundsen@oracle.com $ */
 /** @file
  * VMM testcase - Helper stuff.
  */
@@ -37,7 +37,7 @@ RT_C_DECLS_END
         if (off != RT_OFFSETOF(type, m)) \
         { \
             printf("error! %#010x %s  Off by %d!! (expected off=%#x)\n", \
-                   RT_OFFSETOF(type, m), #type "." #m, off - RT_OFFSETOF(type, m), off); \
+                   RT_OFFSETOF(type, m), #type "." #m, off - RT_OFFSETOF(type, m), (int)off); \
             rc++; \
         } \
         /*else */ \
